requirements.txt
aniso8601==8.0.0
astroid==2.3.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.1
colorama==0.4.3
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
gunicorn==20.0.4
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.6
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
requests==2.24.0
six==1.14.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.11.2
